Changed to a basic React app with TypeScript
