+++ README.md
@@ -1,7 +1,9 @@
-# Discord Activity: Getting Started Guide
+# Connect 4 Discord Activity
 
 This template is used in the [Building An Activity](https://discord.com/developers/docs/activities/building-an-activity) tutorial in the Discord Developer Docs.
 
 Read more about building Discord Activities with the Embedded App SDK at [https://discord.com/developers/docs/activities/overview](https://discord.com/developers/docs/activities/overview).
 
-Updated the README for a test
+## Getting Started
+
+1. Create a .env file and add your client id and secret to it.
